+++ Demo/Sources/Demos/Demo Screens/AutoScrollingViewController.swift
@@ -57,6 +57,7 @@ final class AutoScrollingViewController : UIViewController
     }
 
     private func updateItems(autoScrollIfVisible lastItem : AnyIdentifier? = nil) {
+        
         self.list.configure { list in
             list.appearance = .demoAppearance
             list.layout = .demoLayout