+++ ListableUI/Sources/HeaderFooter/HeaderFooter.swift
@@ -44,7 +44,9 @@ public struct HeaderFooter<Content:HeaderFooterContent> : AnyHeaderFooter
         sizing : Sizing = .thatFits(.init(.atLeast(.default))),
         layout : HeaderFooterLayout = HeaderFooterLayout(),
         onTap : OnTap? = nil
-    ) {        
+    ) {
+        assertIsValueType(Content.self)
+        
         self.content = content
         
         self.sizing = sizing