+++ ListableUI/Sources/Item/Item.swift
@@ -69,7 +69,9 @@ public struct Item<Content:ItemContent> : AnyItem
         onRemove : OnRemove.Callback? = nil,
         onMove : OnMove.Callback? = nil,
         onUpdate : OnUpdate.Callback? = nil
-    ) {        
+    ) {
+        assertIsValueType(Content.self)
+        
         self.content = content
                 
         if let sizing = sizing {